requirements.txt
openai>=1.0.0
requests>=2.31.0
pyyaml>=6.0
pytrends>=4.9.0
praw>=7.7.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
Pillow>=10.0.0
moviepy>=1.0.3
elevenlabs>=0.2.0
gtts>=2.4.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
schedule>=1.2.0
tiktoken>=0.5.0
